reorganizo la estructura de clases para separar mejor el modelo de datos y los componentes de la vista.
