+++ src/components/Presentacion.tsx
@@ -1,4 +1,4 @@
-import { Flyer } from "../model/Flyer";
+import { Flyer } from "./Flyer";
 
 export default function Presentacion( {flyer}: {flyer: Flyer}) {
