--- src/model/IPresentable.tsx
@@ -1,5 +1,0 @@
-import { ReactNode } from "react";
-
-export interface IPresentable {
-    presentacion(): ReactNode;
-}
