+++ src/pages/index.tsx
@@ -5,7 +5,7 @@ import PanelControl from "@/components/PanelControl"
 import Formulario from "@/components/Formulario";
 import Presentacion from "@/components/Presentacion";
 import { useState } from "react";
-import { Flyer } from "@/model/Flyer";
+import { Flyer } from "@/components/Flyer";
 
 const inter = Inter({ subsets: ["latin"] });
